pyproject.toml
[project]
name = "nonebot-plugin-mute-cat"
version = "1.3.0"
description = "The Betterest Mute Cat — 极致的禁言猫猫，功能强大的QQ群禁言插件"
readme = "README.md"
requires-python = ">=3.9,<4.0"
authors = [
    {name = "binglang", email = "lianbingyu_v2@163.com"},
]
license = "MIT"
license-files = ["LICENSE"]
keywords = ["nonebot", "nonebot2", "qq", "mute", "ban", "禁言"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "Natural Language :: Chinese (Simplified)",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Communications :: Chat",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "nonebot2>=2.2.0",
    "nonebot-adapter-onebot>=2.4.0",
    "nonebot-plugin-apscheduler>=0.4.0",
    "nonebot-plugin-localstore>=0.6.0",
    "pydantic>=1.10,<3.0",
]

[project.urls]
homepage = "https://github.com/binglang001/nonebot-plugin-mute-cat"
repository = "https://github.com/binglang001/nonebot-plugin-mute-cat"
documentation = "https://github.com/binglang001/nonebot-plugin-mute-cat#readme"

[build-system]
requires = ["setuptools>=77.0.3"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["nonebot_plugin_mute_cat*"]

[tool.nonebot]
adapters = [{name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11"}]
plugins = ["nonebot_plugin_mute_cat"]
